TexDialog: Add tex command output tab
Previously, it could be difficult to diagnose errors in large equations.
This commit adds a tab to the TeX dialog with the LaTeX generation
command's output.

+++ src/core/control/LatexController.cpp
@@ -113,12 +113,12 @@ void LatexController::findSelectedTexElement() {
 
         if (layout->getPageViewAt(centerX, centerY) == this->view) {
             // Pick the center of the visible area (converting from screen to page coordinates)
-            this->posx = (centerX - this->view->getX()) / zoom;
-            this->posy = (centerY - this->view->getY()) / zoom;
+            this->posx = static_cast<int>((centerX - this->view->getX()) / zoom);
+            this->posy = static_cast<int>((centerY - this->view->getY()) / zoom);
         } else {
             // No better location, so just center it on the page (possibly out of viewport)
-            this->posx = 0.5 * this->page->getWidth();
-            this->posy = 0.5 * this->page->getHeight();
+            this->posx = static_cast<int>(this->page->getWidth() / 2);
+            this->posy = static_cast<int>(this->page->getHeight() / 2);
         }
     }
     this->doc->unlock();
@@ -156,9 +156,12 @@ void LatexController::triggerImageUpdate(const string& texString) {
     if (auto* err = std::get_if<LatexGenerator::GenError>(&result)) {
         XojMsgBox::showErrorToUser(this->control->getGtkWindow(), err->message);
     } else if (auto** proc = std::get_if<GSubprocess*>(&result)) {
+        // Render the TeX and capture the process' output.
         updating_cancellable = g_cancellable_new();
-        g_subprocess_wait_check_async(*proc, updating_cancellable,
-                                      reinterpret_cast<GAsyncReadyCallback>(onPdfRenderComplete), this);
+        char* stdinBuff = nullptr;  // No stdin
+
+        g_subprocess_communicate_utf8_async(*proc, stdinBuff, updating_cancellable,
+                                            reinterpret_cast<GAsyncReadyCallback>(onPdfRenderComplete), this);
     }
 
     updateStatus();
@@ -178,8 +181,24 @@ void LatexController::handleTexChanged(GtkTextBuffer* buffer, LatexController* s
 
 void LatexController::onPdfRenderComplete(GObject* procObj, GAsyncResult* res, LatexController* self) {
     GError* err = nullptr;
+    bool procExited = false;
     GSubprocess* proc = G_SUBPROCESS(procObj);
-    g_subprocess_wait_check_finish(proc, res, &err);
+
+    // Extract the process' output and store it.
+    {
+        char* procStdout_ptr = nullptr;
+
+        // Stdout and stderr should be merged.
+        procExited = g_subprocess_communicate_utf8_finish(proc, res, &procStdout_ptr, nullptr, &err);
+
+        // If we have stdout, store it.
+        if (procStdout_ptr != nullptr) {
+            self->texProcessOutput = procStdout_ptr;
+            free(procStdout_ptr);
+        } else {
+            g_warning("latex command: no stdout stream");
+        }
+    }
 
     if (err != nullptr) {
         if (g_error_matches(err, G_IO_ERROR, G_IO_ERROR_CANCELLED)) {
@@ -193,16 +212,26 @@ void LatexController::onPdfRenderComplete(GObject* procObj, GAsyncResult* res, L
             g_warning("latex: %s", message.c_str());
             XojMsgBox::showErrorToUser(self->control->getGtkWindow(), message);
         }
+
         self->isValidTex = false;
+        g_error_free(err);
+    } else if (procExited && g_subprocess_get_exit_status(proc) != 0) {
+        // Command exited with non-zero exit status.
+
+        self->isValidTex = false;
+    } else {
+        self->isValidTex = true;
+    }
+
+    // Delete the PDF if the TeX is invalid.
+    if (!self->isValidTex) {
         fs::path pdfPath = self->texTmpDir / "tex.pdf";
         fs::remove(pdfPath);
-        g_error_free(err);
     }
 
     const string currentTex = self->dlg.getBufferContents();
     bool shouldUpdate = self->lastPreviewedTex != currentTex;
-    if (err == nullptr) {
-        self->isValidTex = true;
+    if (self->isValidTex) {
         self->temporaryRender = self->loadRendered(currentTex);
         if (self->temporaryRender != nullptr) {
             self->dlg.setTempRender(self->temporaryRender->getPdf());
@@ -236,8 +265,14 @@ void LatexController::updateStatus() {
 
     gtk_widget_set_sensitive(okButton, buttonEnabled);
 
+    // Show error warning only if LaTeX is invalid.
     GtkLabel* errorLabel = GTK_LABEL(this->dlg.get("texErrorLabel"));
     gtk_label_set_text(errorLabel, this->isValidTex ? "" : N_("The formula is empty when rendered or invalid."));
+
+    // Update the output pane.
+    GtkTextView* commandOutputDisplay = GTK_TEXT_VIEW(this->dlg.get("texCommandOutputText"));
+    GtkTextBuffer* commandOutputBuffer = gtk_text_view_get_buffer(commandOutputDisplay);
+    gtk_text_buffer_set_text(commandOutputBuffer, this->texProcessOutput.c_str(), -1);
 }
 
 void LatexController::deleteOldImage() {
